Init submodules in autogen.sh

+++ autogen.sh
@@ -1,6 +1,22 @@
 #!/bin/sh
 set -x
 
+# Make sure we have common
+if test ! -f common/flumotion-trial;
+  then
+  echo "+ Setting up common submodule"
+  git submodule init
+fi
+git submodule update
+
+# source helper functions
+if test ! -f common/flumotion-trial;
+  then
+  echo There is something wrong with your source tree.
+  echo You are missing common/flumotion-trial
+  exit 1
+fi
+
 rm -f .version
 libtoolize -c --force || exit 1
 autopoint --force || exit 1
